Table Data Structure
The table data structure in OpenDB is inspired by the highly regarded table structure in MATLAB, offering a powerful and intuitive way to store and manipulate data. It consists of a javascript object where each key represents a column or row variable, with its corresponding value being an array of equal length. As a fundamental building block of OpenDB, the table serves as a reliable and efficient tool for data management. In the future, we envision the table becoming a formidable competitor to the javascript object itself. 😎